CoD_Lite/requirements.txt
lightning==2.5.0.post0
torch==2.5.0
torchvision==0.20.0
jsonargparse[signatures]>=4.27.7
einops
omegaconf
timm
numpy
Pillow
tqdm
peft
lpips
DISTS_pytorch
torchmetrics
pandas
PyYAML
